pyproject.toml
[project]
name = "image-to-stl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy-stl==2.16.3",
    "numpy>=2.1.3",
    "opencv-contrib-python>=4.10.0.84",
    "opencv-python>=4.10.0.84",
    "pydantic>=2.10.2",
    "pyyaml>=6.0.2",
    "stl>=0.0.3",
]
